requirements.txt
pandas>=2.0.0
tiktoken>=0.6.0
rich>=13.7.0
tenacity>=8.2.0
orjson>=3.9.0
pydantic>=2.5.0
html5lib>=1.1
python-dotenv
tavily-python
langchain
beautifulsoup4
requests
langchain-core
langgraph
langchain-mistralai
langchain-chroma
langchain-huggingface
fastapi
uvicorn